requirements.txt
python-dotenv
langchain-community
streamlit
openai
langchain-openai
chromadb==0.5.3
mysql-connector-python
chromadb